fix: fix windows exceptions on loading and deleting spaces
Loading a space or trying to remove it if the process fails was causing
an EPERM or an ENOTEMPTY exception due to issues with Windows and how
it does not release file handles properly. We are now handling this
using wrappers around the functions so that they are tried multiple
times or wait until a file access is finished.

closes #159